+++ android/src/main/java/io/parity/signer/bottomsheets/NetworkSelector.kt
@@ -1,4 +1,4 @@
-package io.parity.signer.modals
+package io.parity.signer.bottomsheets
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
