+++ android/src/main/java/io/parity/signer/bottomsheets/NewSeedBackup.kt
@@ -1,4 +1,4 @@
-package io.parity.signer.modals
+package io.parity.signer.bottomsheets
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.MaterialTheme
