+++ android/src/main/java/io/parity/signer/bottomsheets/NewSeedMenu.kt
@@ -1,31 +1,26 @@
-package io.parity.signer.modals
+package io.parity.signer.bottomsheets
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
-import androidx.compose.runtime.*
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.State
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import io.parity.signer.alerts.AndroidCalledConfirm
 import io.parity.signer.components.BigButton
 import io.parity.signer.components.HeaderBar
 import io.parity.signer.models.AlertState
 import io.parity.signer.ui.theme.Bg000
 import io.parity.signer.ui.theme.modal
 import io.parity.signer.uniffi.Action
-import io.parity.signer.uniffi.MSeedMenu
 
 @Composable
-fun SeedMenu(
-	seedMenu: MSeedMenu,
+fun NewSeedMenu(
 	alertState: State<AlertState?>,
-	button: (Action) -> Unit,
-	removeSeed: (String) -> Unit
+	button: (Action) -> Unit
 ) {
-	var confirm by remember { mutableStateOf(false) }
-
 	Column {
 		Spacer(Modifier.weight(1f))
 		Surface(
@@ -35,51 +30,26 @@ fun SeedMenu(
 			Column(
 				modifier = Modifier.padding(20.dp)
 			) {
-				HeaderBar(line1 = "SEED MENU", line2 = "Select action")
+				HeaderBar(line1 = "ADD SEED", line2 = "Select seed addition method")
 				BigButton(
-					text = "Backup",
+					text = "New seed",
 					action = {
 						if (alertState.value == AlertState.None)
-							button(Action.BACKUP_SEED)
+							button(Action.NEW_SEED)
 						else
 							button(Action.SHIELD)
-					}
-				)
+					})
 				BigButton(
-					text = "Derive new key",
+					text = "Recover seed",
 					action = {
 						if (alertState.value == AlertState.None)
-							button(Action.NEW_KEY)
+							button(Action.RECOVER_SEED)
 						else
 							button(Action.SHIELD)
 					},
-					isShaded = true,
-					isCrypto = true
-				)
-				BigButton(
-					text = "Forget this seed forever",
-					isShaded = true,
-					isDangerous = true,
-					action = {
-						confirm = true
-					}
+					isShaded = true
 				)
 			}
 		}
 	}
-
-	AndroidCalledConfirm(
-		show = confirm,
-		header = "Forget this seed forever?",
-		text = "This seed will be removed for all networks. " +
-			"This is not reversible. Are you sure?",
-		back = { confirm = false },
-		forward = {
-			seedMenu.seed.let {
-				if (seedMenu.seed.isNotBlank()) removeSeed(it)
-			}
-		},
-		backText = "Cancel",
-		forwardText = "Remove seed"
-	)
 }
