+++ android/src/main/java/io/parity/signer/bottomsheets/SeedMenu.kt
@@ -1,4 +1,4 @@
-package io.parity.signer.modals
+package io.parity.signer.bottomsheets
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
@@ -11,14 +11,19 @@ import androidx.compose.ui.unit.dp
 import io.parity.signer.alerts.AndroidCalledConfirm
 import io.parity.signer.components.BigButton
 import io.parity.signer.components.HeaderBar
-import io.parity.signer.models.SignerDataModel
-import io.parity.signer.models.pushButton
+import io.parity.signer.models.AlertState
 import io.parity.signer.ui.theme.Bg000
 import io.parity.signer.ui.theme.modal
 import io.parity.signer.uniffi.Action
+import io.parity.signer.uniffi.MSeedMenu
 
 @Composable
-fun NetworkDetailsMenu(signerDataModel: SignerDataModel) {
+fun SeedMenu(
+	seedMenu: MSeedMenu,
+	alertState: State<AlertState?>,
+	button: (Action) -> Unit,
+	removeSeed: (String) -> Unit
+) {
 	var confirm by remember { mutableStateOf(false) }
 
 	Column {
@@ -30,14 +35,29 @@ fun NetworkDetailsMenu(signerDataModel: SignerDataModel) {
 			Column(
 				modifier = Modifier.padding(20.dp)
 			) {
-				HeaderBar(line1 = "MANAGE NETWORK", line2 = "Select action")
+				HeaderBar(line1 = "SEED MENU", line2 = "Select action")
 				BigButton(
-					text = "Sign network specs",
+					text = "Backup",
+					action = {
+						if (alertState.value == AlertState.None)
+							button(Action.BACKUP_SEED)
+						else
+							button(Action.SHIELD)
+					}
+				)
+				BigButton(
+					text = "Derive new key",
+					action = {
+						if (alertState.value == AlertState.None)
+							button(Action.NEW_KEY)
+						else
+							button(Action.SHIELD)
+					},
 					isShaded = true,
-					isCrypto = true,
-					action = { signerDataModel.pushButton(Action.SIGN_NETWORK_SPECS) })
+					isCrypto = true
+				)
 				BigButton(
-					text = "Delete network",
+					text = "Forget this seed forever",
 					isShaded = true,
 					isDangerous = true,
 					action = {
@@ -47,13 +67,19 @@ fun NetworkDetailsMenu(signerDataModel: SignerDataModel) {
 			}
 		}
 	}
+
 	AndroidCalledConfirm(
 		show = confirm,
-		header = "Remove network?",
-		text = "This network will be removed for whole device",
+		header = "Forget this seed forever?",
+		text = "This seed will be removed for all networks. " +
+			"This is not reversible. Are you sure?",
 		back = { confirm = false },
-		forward = { signerDataModel.pushButton(Action.REMOVE_NETWORK) },
+		forward = {
+			seedMenu.seed.let {
+				if (seedMenu.seed.isNotBlank()) removeSeed(it)
+			}
+		},
 		backText = "Cancel",
-		forwardText = "Remove network"
+		forwardText = "Remove seed"
 	)
 }
