+++ android/src/main/java/io/parity/signer/models/Keys.kt
@@ -13,7 +13,7 @@ fun SignerDataModel.addKey(path: String, seedName: String) {
 		try {
 			val seedPhrase = getSeed(seedName)
 			if (seedPhrase.isNotBlank()) {
-				pushButton(Action.GO_FORWARD, path, seedPhrase)
+				navigate(Action.GO_FORWARD, path, seedPhrase)
 			}
 		} catch (e: java.lang.Exception) {
 			Log.e("Add key error", e.toString())
@@ -26,7 +26,7 @@ fun SignerDataModel.increment(number: Int, seedName: String) {
 		try {
 			val seedPhrase = getSeed(seedName)
 			if (seedPhrase.isNotBlank()) {
-				pushButton(Action.INCREMENT, number.toString())
+				navigate(Action.INCREMENT, number.toString())
 			}
 		} catch (e: java.lang.Exception) {
 			Log.e("Add key error", e.toString())
