+++ android/src/main/java/io/parity/signer/models/SignerDataModel.kt
@@ -302,7 +302,7 @@ class SignerDataModel : ViewModel() {
 			getAlertState()
 			isAirplaneOn()
 			refreshSeedNames(init = true)
-			pushButton(Action.START)
+			navigate(Action.START)
 		}
 	}
